+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This file records changes to the codebase grouped by version release. Unreleased
 - Improve performance by caching dataframe access during model building
 - Improve performance of junction balancing by adding a scalar fast-path for non-duration-group residual junctions (significant for junction-heavy models)
 - Improve performance by implementing fast-path population aggregations for special cases of no weights, and 1x1 interactions
+- Improve performance by implementing a scalar fast-path for `sdiv` in the function parser
 - Add an MCP server to enable enhanced AI functionality
 
 ## [1.31.7] - 2026-05-29